+++ src/Drivers/Cache/Authorization.php
@@ -1,0 +1,32 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * @author enea dhack <hello@enea.io>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Enea\Authorization\Drivers\Cache;
+
+class Authorization
+{
+    private $name;
+
+    public function __construct(string $name)
+    {
+        $this->name = $name;
+    }
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    public static function make(string $name): Authorization
+    {
+        return new static($name);
+    }
+}