+++ src/Drivers/Database/Authorizer.php
@@ -3,17 +3,20 @@
 declare(strict_types=1);
 
 /**
- * Created on 07/03/18 by enea dhack.
+ * @author enea dhack <hello@enea.io>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
  */
 
 namespace Enea\Authorization\Drivers\Database;
 
-use Enea\Authorization\Authorizer as AuthorizerContract;
 use Enea\Authorization\Contracts\{
     PermissionsOwner, RolesOwner
 };
+use Enea\Authorization\Drivers\Authorizer as BaseAuthorizer;
 
-class Authorizer implements AuthorizerContract
+class Authorizer extends BaseAuthorizer
 {
     private $permission;
 
@@ -25,19 +28,9 @@ class Authorizer implements AuthorizerContract
         $this->role = $role;
     }
 
-    public function can(PermissionsOwner $owner, string $permission): bool
-    {
-        return $this->canAny($owner, [$permission]);
-    }
-
     public function canAny(PermissionsOwner $owner, array $permissions): bool
     {
         return $this->permission->evaluate($owner, $permissions);
-    }
-
-    public function is(RolesOwner $owner, string $role): bool
-    {
-        return $this->isAny($owner, [$role]);
     }
 
     public function isAny(RolesOwner $owner, array $roles): bool