+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 All notable changes to `laravel-authorization` will be documented in this file
+## V0.1.0 - 2018-03-29
+
+### Added
+- Cache driver
+
+### Changed
+- Default driver is now cache
 
 ## V0.0.1 - 2018-03-25